# create-a-container/openapi.v1.yaml
openapi: 3.0.3
info:
  title: Create-a-Container API v1
  description: |
    JSON-only REST API for the React SPA and external automation.
    Co-exists with the legacy `/` (HTML) and `/apikeys` (mixed) routes during the React migration.

    ## Authentication

    Two mechanisms are supported:

    - **Session cookie** — established by `POST /api/v1/auth/login`. Requires a CSRF token (see below) on every state-changing request.
    - **API key** — `Authorization: Bearer <key>`. CSRF is skipped for Bearer-authenticated requests.

    ## CSRF (session auth only)

    1. `GET /api/v1/csrf-token` → returns `{ "data": { "csrfToken": "..." } }` and sets `__Host-csrf.token` cookie.
    2. Send the token in the `X-CSRF-Token` header on every `POST`/`PUT`/`PATCH`/`DELETE`.

    ## Response envelope

    Success: `{ "data": <payload>, "meta"?: {...} }`.
    Error:   `{ "error": { "code": "...", "message": "...", "fields"?: {...} } }`.
  version: 1.0.0
servers:
  - url: /api/v1
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    SessionCookie:
      type: apiKey
      in: cookie
      name: connect.sid
...
security:
  - BearerAuth: []
  - SessionCookie: []
paths:
...
  /sites/{siteId}/nodes:
    parameters: [{ in: path, name: siteId, required: true, schema: { type: integer } }]
    get: { tags: [Nodes], responses: { '200': { description: Array of nodes } } }
    post: { tags: [Nodes], responses: { '201': { description: Created (admin) } } }
  /sites/{siteId}/nodes/import:
    post:
      tags: [Nodes]
      summary: Bulk-import nodes + containers from a Proxmox cluster (admin)
      parameters: [{ in: path, name: siteId, required: true, schema: { type: integer } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [apiUrl, username, password]
              properties:
                apiUrl: { type: string, format: uri }
                username: { type: string }
                password: { type: string }
                tlsVerify: { type: boolean }
      responses: { '201': { description: Imported } }
  /sites/{siteId}/nodes/{id}:
    parameters:
      - { in: path, name: siteId, required: true, schema: { type: integer } }
      - { in: path, name: id, required: true, schema: { type: integer } }
    get: { tags: [Nodes], responses: { '200': { description: Node } } }
    put: { tags: [Nodes], responses: { '200': { description: Updated (admin) } } }
    delete: { tags: [Nodes], responses: { '204': { description: Deleted (admin) } } }
  /sites/{siteId}/nodes/{id}/storages:
    get:
      tags: [Nodes]
      parameters:
        - { in: path, name: siteId, required: true, schema: { type: integer } }
        - { in: path, name: id, required: true, schema: { type: integer } }
      responses: { '200': { description: Array of Proxmox CT template storages } }
  /jobs/{id}:
    parameters: [{ in: path, name: id, required: true, schema: { type: integer } }]
    get: { tags: [Jobs], responses: { '200': { description: Job metadata } } }
  /jobs/{id}/status:
    get:
      tags: [Jobs]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
        - { in: query, name: offset, schema: { type: integer } }
        - { in: query, name: limit, schema: { type: integer } }
      responses: { '200': { description: Status rows } }
  /jobs/{id}/stream:
    get:
      tags: [Jobs]
      summary: Server-sent events stream of job output
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
        - { in: query, name: lastId, schema: { type: integer } }
      responses:
        '200':
          description: SSE stream — events `log` and `status`
          content: { text/event-stream: {} }
...
  /users/invite:
    post:
      tags: [Users]
      summary: Send an email invitation (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties: { email: { type: string, format: email } }
      responses: { '200': { description: Invitation sent } }
  /users/{uidNumber}:
    parameters: [{ in: path, name: uidNumber, required: true, schema: { type: integer } }]
    get: { tags: [Users], responses: { '200': { description: User } } }
    put: { tags: [Users], responses: { '200': { description: Updated (admin) } } }
    delete: { tags: [Users], responses: { '204': { description: Deleted (admin) } } }
...
  /groups/{id}:
    parameters: [{ in: path, name: id, required: true, schema: { type: integer } }]
    get: { tags: [Groups], responses: { '200': { description: Group } } }
    put: { tags: [Groups], responses: { '200': { description: Updated (admin) } } }
    delete: { tags: [Groups], responses: { '204': { description: Deleted (admin) } } }
...